# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EChargers",
    "description": "Echargers API\n",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://52.149.199.210"
    }
  ],
...
  "paths": {
    "/chargers": {
      "get": {
        "tags": [
          "Charger"
        ],
        "summary": "Get all chargers",
        "operationId": "getChargers",
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "500": {
            "description": "Unavailable"
          }
        }
      }
    },
    "/chargers/": {
      "post": {
        "tags": [
          "Charger"
        ],
        "summary": "Create a charger",
        "operationId": "postChargers",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChargerCreate"
        },
        "responses": {
          "200": {
            "description": "ok"
          },
          "500": {
            "description": "Unavailable"
          }
        }
      }
    },
    "/chargers/{chargerId}": {
      "get": {
        "tags": [
          "Charger"
        ],
        "summary": "Find charger by Id",
        "operationId": "getChargerById",
        "parameters": [
          {
            "name": "chargerId",
            "in": "path",
            "description": "ID of charger to return",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Charger"
                }
              }
            }
          },
          "500": {
            "description": "Unavailable"
          }
        }
      },
      "put": {
        "tags": [
          "Charger"
        ],
        "summary": "Update charger",
        "operationId": "putCharger",
        "parameters": [
          {
            "name": "chargerId",
            "in": "path",
            "description": "ID of charger to return",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ChargerUpdate"
        },
        "responses": {
          "200": {
            "description": "ok"
          },
          "500": {
            "description": "Unavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Charger"
        ],
        "summary": "Delete charger",
        "operationId": "deleteCharger",
        "parameters": [
          {
            "name": "chargerId",
            "in": "path",
            "description": "ID of charger to return",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "500": {
            "description": "Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Charger": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "charger1"
          },
          "location": {
            "$ref": "#/components/schemas/Charger_location"
          },
          "averageRating": {
            "type": "number"
          },
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rating"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            }
          },
          "closestCity": {
            "type": "string"
          },
          "temperature": {
            "type": "number"
          },
          "weather": {
            "$ref": "#/components/schemas/Charger_weather"
          }
        }
      },
...
      "Rating": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chargerID": {
            "type": "string"
          },
          "userID": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chargerID": {
            "type": "string"
          },
          "userID": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        }
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chargerID": {
            "type": "string"
          },
          "userID": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        }
      },
      "inline_response_200": {
        "type": "object",
        "properties": {
          "chargers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Charger"
            }
          }
        }
      },
...
      "Charger_location": {
        "type": "object",
        "properties": {
          "longitude": {
            "type": "number",
            "format": "float64"
          },
          "latitude": {
            "type": "number",
            "format": "float64"
          }
        }
      },
      "Charger_weather": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          }
        }
      }
    },
    "requestBodies": {
      "ChargerUpdate": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "averageRating": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
...
      "ChargerCreate": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "longitude": {
                      "type": "number",
                      "format": "float64"
                    },
                    "latitude": {
                      "type": "number",
                      "format": "float64"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWTAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}